Modified verify.ejs to go be authenticate.ejs in the Webserver, changed to server and serveroptimized
